+++ src/platform_impl/macos/window.rs
@@ -30,7 +30,7 @@ use crate::{
     },
     window::{
         CursorGrabMode, CursorIcon, Theme, UserAttentionType, WindowAttributes, WindowButtons,
-        WindowId as RootWindowId,
+        WindowId as RootWindowId, WindowLevel,
     },
 };
 use core_graphics::display::{CGDisplay, CGPoint};
@@ -45,7 +45,8 @@ use objc2::{declare_class, msg_send, msg_send_id, sel, ClassType};
 use super::appkit::{
     NSApp, NSAppKitVersion, NSAppearance, NSApplicationPresentationOptions, NSBackingStoreType,
     NSColor, NSCursor, NSFilenamesPboardType, NSRequestUserAttentionType, NSResponder, NSScreen,
-    NSWindow, NSWindowButton, NSWindowLevel, NSWindowStyleMask, NSWindowTitleVisibility,
+    NSWindow, NSWindowButton, NSWindowLevel, NSWindowSharingType, NSWindowStyleMask,
+    NSWindowTitleVisibility,
 };
 
 #[derive(Debug, Copy, Clone, PartialEq, Eq, PartialOrd, Ord, Hash)]
@@ -313,6 +314,10 @@ impl WinitWindow {
                 this.setTitle(&NSString::from_str(&attrs.title));
                 this.setAcceptsMouseMovedEvents(true);
 
+                if attrs.content_protected {
+                    this.setSharingType(NSWindowSharingType::NSWindowSharingNone);
+                }
+
                 if pl_attrs.titlebar_transparent {
                     this.setTitlebarAppearsTransparent(true);
                 }
@@ -334,10 +339,6 @@ impl WinitWindow {
                 }
                 if pl_attrs.movable_by_window_background {
                     this.setMovableByWindowBackground(true);
-                }
-
-                if attrs.always_on_top {
-                    this.setLevel(NSWindowLevel::Floating);
                 }
 
                 if !attrs.enabled_buttons.contains(WindowButtons::MAXIMIZE) {
@@ -402,6 +403,8 @@ impl WinitWindow {
         if let Some(dim) = attrs.max_inner_size {
             this.set_max_inner_size(Some(dim));
         }
+
+        this.set_window_level(attrs.window_level);
 
         // register for drag and drop operations.
         this.registerForDraggedTypes(&NSArray::from_slice(&[
@@ -795,7 +798,7 @@ impl WinitWindow {
         shared_state_lock.fullscreen = None;
 
         let maximized = shared_state_lock.maximized;
-        let mask = self.saved_style(&mut *shared_state_lock);
+        let mask = self.saved_style(&mut shared_state_lock);
 
         drop(shared_state_lock);
 
@@ -1075,11 +1078,11 @@ impl WinitWindow {
     }
 
     #[inline]
-    pub fn set_always_on_top(&self, always_on_top: bool) {
-        let level = if always_on_top {
-            NSWindowLevel::Floating
-        } else {
-            NSWindowLevel::Normal
+    pub fn set_window_level(&self, level: WindowLevel) {
+        let level = match level {
+            WindowLevel::AlwaysOnTop => NSWindowLevel::Floating,
+            WindowLevel::AlwaysOnBottom => NSWindowLevel::BELOW_NORMAL,
+            WindowLevel::Normal => NSWindowLevel::Normal,
         };
         util::set_level_async(self, level);
     }
@@ -1184,6 +1187,14 @@ impl WinitWindow {
     }
 
     #[inline]
+    pub fn set_content_protected(&self, protected: bool) {
+        self.setSharingType(if protected {
+            NSWindowSharingType::NSWindowSharingNone
+        } else {
+            NSWindowSharingType::NSWindowSharingReadOnly
+        })
+    }
+
     pub fn title(&self) -> String {
         self.title_().to_string()
     }
@@ -1252,7 +1263,7 @@ impl WindowExtMacOS for WinitWindow {
 
             true
         } else {
-            let new_mask = self.saved_style(&mut *shared_state_lock);
+            let new_mask = self.saved_style(&mut shared_state_lock);
             self.set_style_mask_async(new_mask);
             shared_state_lock.is_simple_fullscreen = false;
 
@@ -1276,6 +1287,14 @@ impl WindowExtMacOS for WinitWindow {
     #[inline]
     fn set_has_shadow(&self, has_shadow: bool) {
         self.setHasShadow(has_shadow)
+    }
+
+    fn is_document_edited(&self) -> bool {
+        self.isDocumentEdited()
+    }
+
+    fn set_document_edited(&self, edited: bool) {
+        self.setDocumentEdited(edited)
     }
 }