+++ src/platform_impl/web/window.rs
@@ -4,7 +4,7 @@ use crate::event;
 use crate::icon::Icon;
 use crate::window::{
     CursorGrabMode, CursorIcon, Theme, UserAttentionType, WindowAttributes, WindowButtons,
-    WindowId as RootWI,
+    WindowId as RootWI, WindowLevel,
 };
 
 use raw_window_handle::{RawDisplayHandle, RawWindowHandle, WebDisplayHandle, WebWindowHandle};
@@ -258,7 +258,7 @@ impl Window {
         } else {
             self.canvas
                 .borrow()
-                .set_attribute("cursor", *self.previous_pointer.borrow());
+                .set_attribute("cursor", &self.previous_pointer.borrow());
         }
     }
 
@@ -316,7 +316,7 @@ impl Window {
     }
 
     #[inline]
-    pub fn set_always_on_top(&self, _always_on_top: bool) {
+    pub fn set_window_level(&self, _level: WindowLevel) {
         // Intentionally a no-op, no window ordering
     }