+++ src/platform_impl/windows/window_state.rs
@@ -12,15 +12,15 @@ use windows_sys::Win32::{
     Graphics::Gdi::InvalidateRgn,
     UI::WindowsAndMessaging::{
         AdjustWindowRectEx, EnableMenuItem, GetMenu, GetSystemMenu, GetWindowLongW, SendMessageW,
-        SetWindowLongW, SetWindowPos, ShowWindow, GWL_EXSTYLE, GWL_STYLE, HWND_NOTOPMOST,
-        HWND_TOPMOST, MF_BYCOMMAND, MF_DISABLED, MF_ENABLED, SC_CLOSE, SWP_ASYNCWINDOWPOS,
-        SWP_FRAMECHANGED, SWP_NOACTIVATE, SWP_NOMOVE, SWP_NOREPOSITION, SWP_NOSIZE, SWP_NOZORDER,
-        SW_HIDE, SW_MAXIMIZE, SW_MINIMIZE, SW_RESTORE, SW_SHOW, WINDOWPLACEMENT, WINDOW_EX_STYLE,
-        WINDOW_STYLE, WS_BORDER, WS_CAPTION, WS_CHILD, WS_CLIPCHILDREN, WS_CLIPSIBLINGS,
-        WS_EX_ACCEPTFILES, WS_EX_APPWINDOW, WS_EX_LAYERED, WS_EX_NOREDIRECTIONBITMAP,
-        WS_EX_TOPMOST, WS_EX_TRANSPARENT, WS_EX_WINDOWEDGE, WS_MAXIMIZE, WS_MAXIMIZEBOX,
-        WS_MINIMIZE, WS_MINIMIZEBOX, WS_OVERLAPPEDWINDOW, WS_POPUP, WS_SIZEBOX, WS_SYSMENU,
-        WS_VISIBLE,
+        SetWindowLongW, SetWindowPos, ShowWindow, GWL_EXSTYLE, GWL_STYLE, HWND_BOTTOM,
+        HWND_NOTOPMOST, HWND_TOPMOST, MF_BYCOMMAND, MF_DISABLED, MF_ENABLED, SC_CLOSE,
+        SWP_ASYNCWINDOWPOS, SWP_FRAMECHANGED, SWP_NOACTIVATE, SWP_NOMOVE, SWP_NOREPOSITION,
+        SWP_NOSIZE, SWP_NOZORDER, SW_HIDE, SW_MAXIMIZE, SW_MINIMIZE, SW_RESTORE, SW_SHOW,
+        WINDOWPLACEMENT, WINDOW_EX_STYLE, WINDOW_STYLE, WS_BORDER, WS_CAPTION, WS_CHILD,
+        WS_CLIPCHILDREN, WS_CLIPSIBLINGS, WS_EX_ACCEPTFILES, WS_EX_APPWINDOW, WS_EX_LAYERED,
+        WS_EX_NOREDIRECTIONBITMAP, WS_EX_TOPMOST, WS_EX_TRANSPARENT, WS_EX_WINDOWEDGE, WS_MAXIMIZE,
+        WS_MAXIMIZEBOX, WS_MINIMIZE, WS_MINIMIZEBOX, WS_OVERLAPPEDWINDOW, WS_POPUP, WS_SIZEBOX,
+        WS_SYSMENU, WS_VISIBLE,
     },
 };
 
@@ -77,41 +77,42 @@ bitflags! {
 }
 bitflags! {
     pub struct WindowFlags: u32 {
-        const RESIZABLE      = 1 << 0;
-        const MINIMIZABLE    = 1 << 1;
-        const MAXIMIZABLE    = 1 << 2;
-        const CLOSABLE       = 1 << 3;
-        const VISIBLE        = 1 << 4;
-        const ON_TASKBAR     = 1 << 5;
-        const ALWAYS_ON_TOP  = 1 << 6;
-        const NO_BACK_BUFFER = 1 << 7;
-        const TRANSPARENT    = 1 << 8;
-        const CHILD          = 1 << 9;
-        const MAXIMIZED      = 1 << 10;
-        const POPUP          = 1 << 11;
+        const RESIZABLE         = 1 << 0;
+        const MINIMIZABLE       = 1 << 1;
+        const MAXIMIZABLE       = 1 << 2;
+        const CLOSABLE          = 1 << 3;
+        const VISIBLE           = 1 << 4;
+        const ON_TASKBAR        = 1 << 5;
+        const ALWAYS_ON_TOP     = 1 << 6;
+        const ALWAYS_ON_BOTTOM  = 1 << 7;
+        const NO_BACK_BUFFER    = 1 << 8;
+        const TRANSPARENT       = 1 << 9;
+        const CHILD             = 1 << 10;
+        const MAXIMIZED         = 1 << 11;
+        const POPUP             = 1 << 12;
 
         /// Marker flag for fullscreen. Should always match `WindowState::fullscreen`, but is
         /// included here to make masking easier.
-        const MARKER_EXCLUSIVE_FULLSCREEN = 1 << 12;
-        const MARKER_BORDERLESS_FULLSCREEN = 1 << 13;
+        const MARKER_EXCLUSIVE_FULLSCREEN = 1 << 13;
+        const MARKER_BORDERLESS_FULLSCREEN = 1 << 14;
 
         /// The `WM_SIZE` event contains some parameters that can effect the state of `WindowFlags`.
         /// In most cases, it's okay to let those parameters change the state. However, when we're
         /// running the `WindowFlags::apply_diff` function, we *don't* want those parameters to
         /// effect our stored state, because the purpose of `apply_diff` is to update the actual
         /// window's state to match our stored state. This controls whether to accept those changes.
-        const MARKER_RETAIN_STATE_ON_SIZE = 1 << 14;
+        const MARKER_RETAIN_STATE_ON_SIZE = 1 << 15;
 
-        const MARKER_IN_SIZE_MOVE = 1 << 15;
+        const MARKER_IN_SIZE_MOVE = 1 << 16;
 
-        const MINIMIZED = 1 << 16;
+        const MINIMIZED = 1 << 17;
 
-        const IGNORE_CURSOR_EVENT = 1 << 17;
+        const IGNORE_CURSOR_EVENT = 1 << 18;
 
         /// Fully decorated window (incl. caption, border and drop shadow).
-        const MARKER_DECORATIONS = 1 << 18;
+        const MARKER_DECORATIONS = 1 << 19;
         /// Drop shadow for undecorated windows.
-        const MARKER_UNDECORATED_SHADOW = 1 << 19;
+        const MARKER_UNDECORATED_SHADOW = 1 << 20;
 
         const EXCLUSIVE_FULLSCREEN_OR_MASK = WindowFlags::ALWAYS_ON_TOP.bits;
     }
@@ -306,13 +307,18 @@ impl WindowFlags {
             }
         }
 
-        if diff.contains(WindowFlags::ALWAYS_ON_TOP) {
+        if diff.intersects(WindowFlags::ALWAYS_ON_TOP | WindowFlags::ALWAYS_ON_BOTTOM) {
             unsafe {
                 SetWindowPos(
                     window,
-                    match new.contains(WindowFlags::ALWAYS_ON_TOP) {
-                        true => HWND_TOPMOST,
-                        false => HWND_NOTOPMOST,
+                    match (
+                        new.contains(WindowFlags::ALWAYS_ON_TOP),
+                        new.contains(WindowFlags::ALWAYS_ON_BOTTOM),
+                    ) {
+                        (true, false) => HWND_TOPMOST,
+                        (false, false) => HWND_NOTOPMOST,
+                        (false, true) => HWND_BOTTOM,
+                        (true, true) => unreachable!(),
                     },
                     0,
                     0,